+++ feature/timer/src/main/java/org/sopt/timer/model/Clip.kt
@@ -1,4 +1,4 @@
-package org.sopt.timer.dummymodel
+package org.sopt.timer.model
 
 data class Clip(
   val name: String,
