+++ feature/timer/src/main/java/org/sopt/timer/model/PickerItem.kt
@@ -1,4 +1,4 @@
-package org.sopt.timer.dummymodel
+package org.sopt.timer.model
 
 data class PickerItem(
   val text: String,
