+++ feature/timer/src/main/java/org/sopt/timer/model/TimePicker.kt
@@ -1,4 +1,4 @@
-package org.sopt.timer.dummymodel
+package org.sopt.timer.model
 
 data class TimePicker(
   val timePeriod: String,
